Refactor Region into a struct
Now the Region is a Struct, we can store information inside it that is
common to all kinds of regions, such as the coordinates of the cells
that the region is composed of.

This means we don't have to recompute the coordinates every time we want
to figure out whether a given cell is inside the region or not.