Actually store computed coordinates for regions
We previously computed the coordinates for all the cells for a region
and then promptly forgot all about them again, storing an empty HashSet
in the Region struct's coord member.

Now that this is fixed, we can also rely on the coords HashSet to check
whether a cell is part of the region - we do not need to defer this to
the RegionKind enum.