Implement the region intersection reduction strategy
This introduces the concept of a reduction, which is a coordinate at
which a candidate can be erased. This can be the result of an analysis
like the region intersection one. E.g, if in row 1 a candidate only
occurs in the first three cells, we can conclude that the value 1 cannot
be a candidate for any of the other 6 cells in the top-left square
subgrid.